refactor(#161): userState에서 정보를 받아오도록 수정
